+++ tests/func/test_list.cc
@@ -334,7 +334,7 @@ int main(void)
     int fail = 0;
     INIT_SRAND;
     INIT_TEST_LOOPS(10);
-    for (size_t loop = 0; loop < loops; loop++)
+    for (unsigned loop = 0; loop < loops; loop++)
     {
         list_digi a, aa, aaa;
         std::list<DIGI> b, bb, bbb;
